Renamed individualPosition to positionDescription and moved positionDescription and downloadPositionDescription out of main_routes.py and into departmentPortal.py.
